+++ src/common/store/index.js
@@ -1,9 +1,3 @@
-import { useProgressBarStore } from './progress-bar-steps';
-import { useLayersStore } from './layers.store';
-import { useGeometryStore } from './geometry.store';
-import { useLevelsStore } from './levels.store';
-import { useReviewManifestStore } from './review-manifest.store';
-
 export { getDomain, useUserStore } from './user.store';
 export { useResponseStore } from './response.store';
 export { progressBarSteps, progressBarStepsByKey, useCompletedSteps, useProgressBarStore } from './progress-bar-steps';
@@ -12,11 +6,4 @@ export { useLayersStore } from './layers.store';
 export { useLevelsStore } from './levels.store';
 export { useReviewManifestJson, useReviewManifestStore } from './review-manifest.store';
 export { useConversionStore, conversionSteps } from './conversion.store';
-
-export function resetStores() {
-  useProgressBarStore.getState().reset();
-  useLayersStore.getState().reset();
-  useGeometryStore.getState().reset();
-  useLevelsStore.getState().reset();
-  useReviewManifestStore.getState().setCanBeDownloaded(false);
-}
+export { resetStores } from './reset';
